+++ packages/beff-wasm/bundled-code/build-parsers.js
@@ -57,8 +57,6 @@ function buildParsers(args) {
       parse,
       safeParse,
       zod,
-      //@ts-ignore
-      jsonSchema: jsonSchema[k]
     };
   });
   return decoders;