+++ packages/beff-wasm/fixtures/codegen-snaps/export-decoder/bff-generated/parser.js
@@ -2,7 +2,6 @@
 /* eslint-disable */
 
 
-import jsonSchema from "./schema.js";
 import {printErrors} from '@beff/client';
 import {z} from 'zod';
 import validatorsMod from "./validators.js"; const { decodeObject, decodeArray, decodeString, decodeNumber, decodeCodec, decodeStringWithFormat, decodeAnyOf, decodeAllOf, decodeBoolean, decodeAny, decodeTuple, decodeNull, decodeConst, registerCustomFormatter, validators, c } = validatorsMod;
@@ -93,8 +92,6 @@ function buildParsers(args) {
       parse,
       safeParse,
       zod,
-      
-      jsonSchema: jsonSchema[k]
     };
   });
   return decoders;